+++ src/main/java/camp/woowak/lab/order/domain/vo/OrderItem.java
@@ -7,14 +7,14 @@ import lombok.Getter;
 @Embeddable
 public class OrderItem {
 	private Long menuId;
-	private int price;
+	private long price;
 	private int quantity;
-	private int totalPrice;
+	private long totalPrice;
 
 	protected OrderItem() {
 	}
 
-	public OrderItem(Long menuId, int price, int quantity) {
+	public OrderItem(Long menuId, long price, int quantity) {
 		this.menuId = menuId;
 		this.price = price;
 		this.quantity = quantity;
